[Chore #117945253]Refactor API route handling
