+++ picturethis/imageeditor/urls.py
@@ -1,6 +1,6 @@
 from django.conf.urls import url
 from rest_framework.urlpatterns import format_suffix_patterns
-from views import PhotoListView, EditedPhotoListView
+from viewsets import PhotoListView, EditedPhotoListView
 
 
 urlpatterns = [
